requirements.txt
blessings==1.6
chardet==2.3.0
cookies==2.2.1
docopt==0.6.2
fancycompleter==0.4
ordereddict==1.1
py==1.4.31
pytest==2.9.1
PyYAML==3.11
requests==2.9.1
responses==0.5.1
six==1.10.0
tqdm==3.8.0
wmctrl==0.3
